Added: Check to make sure a crypto key request is only sent once Added: A message with the reason a client is disconnecting Fixed: The spawn info struct is more accurate Changed: The character create packet now reads/saves the byte values for sliders instead of the floats Fixed: loading some more character colors (shirt/pants) and fixed eye color loading

+++ trunk/source/common/Packets/PacketElements/PacketPackedData.h
@@ -88,7 +88,7 @@ public:
 		CheckElementsInitialized();
 
 		if (bBufInitialized) {
-			return buf.size();
+			return static_cast<uint32_t>(buf.size());
 		}
 
 		uint32_t unpackedSize = 0;
